Moved alertIsVisible() method to Base Page and added Header component
